+++ Tests/Performance/UniTask/Task233VsUniTaskPerformanceTests.cs
@@ -1,0 +1,55 @@
+#if TASK233_HAS_UNITASK
+using Cysharp.Threading.Tasks;
+using NUnit.Framework;
+using Unity.PerformanceTesting;
+
+namespace Task233.Tests
+{
+    public sealed class Task233VsUniTaskPerformanceTests
+    {
+        [Test, Performance]
+        public void Task233YieldFactory()
+        {
+            Measure.Method(() => _ = T233.Yield())
+                .WarmupCount(20)
+                .MeasurementCount(100)
+                .IterationsPerMeasurement(10000)
+                .GC()
+                .Run();
+        }
+
+        [Test, Performance]
+        public void UniTaskYieldFactory()
+        {
+            Measure.Method(() => _ = UniTask.Yield())
+                .WarmupCount(20)
+                .MeasurementCount(100)
+                .IterationsPerMeasurement(10000)
+                .GC()
+                .Run();
+        }
+
+        [Test, Performance]
+        public void Task233DelayFrameFactory()
+        {
+            Measure.Method(() => _ = T233.DelayFrames(1))
+                .WarmupCount(20)
+                .MeasurementCount(100)
+                .IterationsPerMeasurement(10000)
+                .GC()
+                .Run();
+        }
+
+        [Test, Performance]
+        public void UniTaskDelayFrameFactory()
+        {
+            Measure.Method(() => _ = UniTask.DelayFrame(1))
+                .WarmupCount(20)
+                .MeasurementCount(100)
+                .IterationsPerMeasurement(10000)
+                .GC()
+                .Run();
+        }
+    }
+}
+#endif